% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_records.R
\name{get_records}
\alias{get_records}
\title{Get Records for Zoho Creator}
\usage{
get_records(account_owner_name, app_name, report_name, access_token)
}
\arguments{
\item{account_owner_name}{Nome do proprietário da conta.}

\item{app_name}{Nome do aplicativo dentro do Zoho Creator.}

\item{report_name}{Nome do relatório de onde os registros serão extraídos.}

\item{access_token}{Token de acesso para autenticação na API do Zoho.}
}
\value{
Um data frame contendo os registros extraídos do relatório.
}
\description{
Esta função busca registros de um relatório no Zoho Creator usando a API.
}
\details{
A função realiza múltiplas requisições à API para recuperar até 1000 registros por vez,
continuando enquanto houver registros adicionais disponíveis. A função só para
depois de buscar todos os registos do relatorio.

#' @examples
}
\section{Exemplo de uso da função com autenticação via arquivo .env}{
library(dotenv)
load_dot_env(".env")

account_owner <- Sys.getenv("ZOHO_ACCOUNT_OWNER")
app <- Sys.getenv("ZOHO_APP_NAME")
report <- Sys.getenv("ZOHO_REPORT_NAME")
token <- Sys.getenv("ZOHO_ACCESS_TOKEN")

registros <- get_records(account_owner, app, report, token)
view(registros)
}

