% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_access_token.R
\name{refresh_access_token}
\alias{refresh_access_token}
\title{Refresh Access Token}
\usage{
refresh_access_token(client_id, client_secret, refresh_token)
}
\arguments{
\item{client_id}{O ID do cliente fornecido pela Zoho.}

\item{client_secret}{O segredo do cliente fornecido pela Zoho.}

\item{refresh_token}{O token de renovação fornecido pela Zoho

Segue https://www.zoho.com/creator/help/api/v2.1/oauth-overview.html Para ter
acesso ao client_id, client_secret, refresh_token.}
}
\value{
Retorna a resposta completa da requisição em formato JSON, que geralmente contém os seguintes campos:
\itemize{
\item \code{access_token}: O novo token de acesso, utilizado para autenticar as requisições à API.
\item \code{expires_in}: O tempo de expiração do access token, geralmente em segundos.
\item \code{token_type}: O tipo do token retornado, normalmente "Bearer".
\item \code{api_domain}: O domínio da API que deve ser usado para fazer chamadas autenticadas.
\item \code{error} (opcional): Caso ocorra um erro na requisição, esse campo pode conter a descrição do erro.
}
}
\description{
Esta função renova o access token utilizando o refresh token fornecido.
}
\examples{
client_id <- "seu_client_id"
client_secret <- "seu_client_secret"
refresh_token <- "seu_refresh_token"
response <- refresh_access_token(client_id, client_secret, refresh_token)

# Exemplo de resposta esperada:
# {
#   "access_token": "1000.abcdef1234567890ghijklmnopqrstu",
#   "refresh_token": "1000.zyxwvutsrqponmlkjihgfedcba123456",
#   "api_domain": "https://www.zohoapis.com",
#   "token_type": "Bearer",
#   "expires_in": 3600
# }
}
